% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog_2025.R
\docType{data}
\name{commune_2025}
\alias{commune_2025}
\title{French Communes Data}
\format{
A data frame with 12 columns:
\describe{
\item{TYPECOM}{(character) Commune type, 4 characters.}
\item{COM}{(character) Commune code, 5 digits.}
\item{REG}{(integer) Region code, 2 digits.}
\item{DEP}{(character) Department code, 3 characters.}
\item{CTCD}{(character) Code of the territorial community with departmental powers, 4 characters.}
\item{ARR}{(character) Borough code, 4 characters.}
\item{TNCC}{(integer) Type of clear name, 1 digit.}
\item{NCC}{(character) Clear name (uppercase), up to 200 characters.}
\item{NCCENR}{(character) Clear name with rich typography, up to 200 characters.}
\item{LIBELLE}{(character) Clear name with rich typography and article, up to 200 characters.}
\item{CAN}{(character) Canton code, 5 characters. For “multi-cantonal” communes, the code ranges from 99 to 90 (pseudo-canton) or from 89 to 80 (new communes).}
\item{COMPARENT}{(integer) Parent commune code for municipal borough and associated or delegated communes, 5 digits.}
}
}
\source{
\url{https://www.insee.fr/fr/information/8377162}
}
\usage{
commune_2025
}
\description{
A dataset containing information about French communes,
including their codes, types, and names.
}
\keyword{datasets}
