% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdg.R
\name{construct_commune}
\alias{construct_commune}
\title{Construct a commune path string}
\usage{
construct_commune(id)
}
\arguments{
\item{id}{\code{character} vector. Validated INSEE code(s) of the commune(s).}
}
\value{
A \code{character} vector combining department and commune codes separated by a slash.
}
\description{
Constructs a path string for a given commune (or sheet) code by adding
the department code (first two characters, or three for 97x)
as a prefix joined with a slash.
}
\examples{
\dontrun{
construct_commune("72187")
# Returns: "72/72187"
construct_commune(c("72187", "75056"))
# Returns: c("72/72187", "75/75056")
}

}
\keyword{internal}
