% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_download.R
\name{extract_file}
\alias{extract_file}
\title{Extract a file (GZ file or archive)}
\usage{
extract_file(filepath, extract_dir)
}
\arguments{
\item{filepath}{Character, path to the file to extract}

\item{extract_dir}{Character, directory where files will be extracted}
}
\value{
Path to extraction directory
}
\description{
Extract a file (GZ file or archive)
}
\keyword{internal}
