% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etalab.R
\name{get_etalab}
\alias{get_etalab}
\title{Download Etalab layer (raw or processed) for one or multiple communes}
\usage{
get_etalab(
  commune,
  layer,
  millesime = "latest",
  extract_dir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{commune}{\code{character} or \code{numeric} vector.
The INSEE code(s) of the commune(s).}

\item{layer}{\code{character}. Datasets raw name to download.
Must be dataset names returned by \code{\link[=get_etalab_layernames]{get_etalab_layernames()}}.}

\item{millesime}{\code{character}. Dataset version for raw layers.
Must be on of \code{get_data_millesimes("etalab")}. Default is \code{"latest"}.}

\item{extract_dir}{\code{character} or \code{NULL}.
Directory where files will be downloaded and extracted.
If \code{NULL}, a temporary directory is used.}

\item{verbose}{\code{logical}. If \code{TRUE}, prints progress messages.}
}
\value{
An \code{sf} object containing the requested layer, or \code{NULL} if download/read fails
}
\description{
This function automatically selects the appropriate download method
based on whether the requested layer is raw or processed.
}
\examples{
\dontrun{
# Download processed layer
batiments <- get_etalab("72187", "batiments")

# Download raw layer
parcelles <- get_etalab("72187", "parcelle")
}

}
