% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etalab.R
\name{get_etalab_raw}
\alias{get_etalab_raw}
\title{Download and read Etalab raw datasets from server}
\usage{
get_etalab_raw(
  id,
  layer,
  millesime = "latest",
  extract_dir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{id}{\code{character} or \code{numeric} vector.
The INSEE code(s) of the commune(s) or department(s).}

\item{layer}{\code{character}. Datasets raw name to download.
Must be dataset names returned by \code{\link[=get_etalab_layernames]{get_etalab_layernames()}}.}

\item{millesime}{\code{character}. Dataset version for raw layers.
Must be on of \code{get_data_millesimes("etalab")}. Default is \code{"latest"}.}

\item{extract_dir}{\code{character} or \code{NULL}.
Directory where files will be downloaded and extracted.
If \code{NULL}, a temporary directory is used.}

\item{verbose}{\code{logical}. If \code{TRUE}, prints progress messages.}
}
\value{
An \code{sf} object containing the requested layer, or \code{NULL} if download/read fails
}
\description{
This function downloads Etalab cadastre dataset for given communes
and layer.
The results are returned as an \code{sf} object. If multiple files are retrieved
(e.g. several communes), they are combined into a single \code{sf} object.
}
\details{
A detailed description of the data structure is available at the following link:
\url{https://github.com/etalab/edigeo-parser/raw/master/resources/standard_edigeo_2013.pdf}
}
\examples{
\dontrun{
# Download and read parcels for one commune
get_etalab_raw("72187", "parcelle")

# Download and read parcels for multiple communes
get_etalab_raw(c("72187", "72181"), "parcelle")
}

}
\seealso{
\code{\link[=get_etalab_layernames]{get_etalab_layernames()}}, \code{\link[=get_data_millesimes]{get_data_millesimes()}}
}
\keyword{internal}
