% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdg.R
\name{get_insee_scale}
\alias{get_insee_scale}
\title{Get administrative scale of INSEE codes}
\usage{
get_insee_scale(x)
}
\arguments{
\item{x}{\code{character} or \code{numeric}. Vector of INSEE codes.}
}
\value{
Character vector of length \code{length(x)} with values "communes",
"departements" or "feuilles".
}
\description{
Returns administrative scale for each INSEE code.
}
\examples{
\dontrun{
get_insee_scale(72187)        # "communes"
get_insee_scale(c(72, 72187)) # c("departements", "communes")
get_insee_scale(c(72, 72187, "721870000A01")) # c("departements", "communes", "feuilles")
}

}
\keyword{internal}
