% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idu.R
\name{idu_get_parcelle}
\alias{idu_get_parcelle}
\title{Retrieve parcel data for given IDUs}
\usage{
idu_get_parcelle(
  idu,
  with_feuille = TRUE,
  with_lieudit = TRUE,
  with_cog = TRUE,
  ...
)
}
\arguments{
\item{idu}{A \code{character} vector of valid IDU codes.}

\item{with_feuille}{\code{logical} (default: \code{TRUE}).
Whether to retrieve and merge sheet id associated with the parcels.}

\item{with_lieudit}{\code{logical} (default: \code{TRUE}).
Whether to retrieve and merge place names associated with the parcels.}

\item{with_cog}{\code{logical} (default: \code{TRUE}).
Whether to retrieve and merge administrative names
(region, department, commune) associated with the parcels.}

\item{...}{Additional arguments passed to \code{\link[=idu_get_cog]{idu_get_cog()}}.}
}
\value{
An \code{sf} object containing parcel geometries, the IDU code, and optionally
associated place names and administrative names.
}
\description{
This function takes one or more valid IDU codes, retrieves parcel data from
Etalab, and optionally enriches the parcels with their associated place
names and administrative names. The function returns an \code{sf} object containing
the parcels and requested attributes.
}
\details{
\itemize{
\item All IDU codes are validated before any data is retrieved.
\item If \code{with_feuille = TRUE}, the function performs a spatial join with the
Etalab "lieux-dits" dataset and merges the names into the parcel data.
\item If \code{with_lieudit = TRUE}, the function performs a spatial join with the
Etalab "lieux-dits" dataset and merges the names into the parcel data.
\item If \code{with_cog = TRUE}, the function retrieves region, department, and commune
names using \code{\link[=idu_get_cog]{idu_get_cog()}} and merges them into the parcel data.
\item The function ensures that Etalab data are returned as \code{sf} objects.
}
}
\examples{
\dontrun{
# Retrieve parcels with both lieudit and names
idu_get_parcelle(c("721870000A0001", "721870000A0002"))

# Retrieve parcels without lieudit
idu_get_parcelle("721870000A0001", with_lieudit = FALSE)

# Retrieve parcels without administrative names
idu_get_parcelle("721870000A0001", with_cog = FALSE)
}

}
