% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etalab.R
\name{get_etalab_proc}
\alias{get_etalab_proc}
\title{Download and read Etalab processed datasets from "bundle" app}
\usage{
get_etalab_proc(id, layer, verbose = TRUE)
}
\arguments{
\item{id}{\code{character} or \code{numeric} vector.
The INSEE code(s) of the commune(s) or department(s).}

\item{layer}{\code{character}. Datasets raw name to download.
Must be dataset names returned by \code{\link[=get_etalab_layernames]{get_etalab_layernames()}}.}

\item{verbose}{\code{logical}. If \code{TRUE}, prints progress messages.}
}
\value{
An \code{sf} object containing the requested layer, or \code{NULL} if download/read fails
}
\description{
This function downloads Etalab cadastral dataset for given communes and
layer, by using the Etalab "bundle" cadastre.data.gouv.
The results are returned as an \code{sf} object. If multiple files are retrieved
(e.g. several communes), they are combined into a single \code{sf} object.
}
\examples{
\dontrun{
# Download parcel geometries for a single commune
get_etalab_proc("72187", data = "parcelles")

# Download parcel geometries for multiple communes
get_etalab_proc(c("72187", "72181"), data = "parcelles")

}

}
\seealso{
\code{\link[=get_etalab_layernames]{get_etalab_layernames()}}
}
\keyword{internal}
